+++ DMRGatewayNetwork.h
@@ -33,7 +33,7 @@
 class CDMRGatewayNetwork : public IDMRNetwork
 {
 public:
-	CDMRGatewayNetwork(const std::string& address, unsigned int port, unsigned int local, unsigned int id, bool duplex, const char* version, bool slot1, bool slot2, HW_TYPE hwType, bool debug);
+	CDMRGatewayNetwork(const std::string& address, unsigned short port, unsigned short local, unsigned int id, bool duplex, const char* version, bool slot1, bool slot2, HW_TYPE hwType, bool debug);
 	virtual ~CDMRGatewayNetwork();
 
 	virtual void setOptions(const std::string& options);
@@ -64,7 +64,7 @@ private:
 	std::string      m_addressStr;
 	sockaddr_storage m_addr;
 	unsigned int     m_addrLen;
-	unsigned int     m_port;
+	unsigned short   m_port;
 	uint8_t*         m_id;
 	bool             m_duplex;
 	const char*      m_version;
